Vs/requirement.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2     
autopep8==2.0.1
beautifulsoup4==4.11.2   
billiard==3.6.4.0        
bs4==0.0.1
celery==5.2.7
certifi==2022.12.7       
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
cryptography==39.0.1
DateTime==5.0
Deprecated==1.2.13
distlib==0.3.6
Django==4.1.6
django-cors-headers==3.14.0
django-crispy-forms==1.14.0
django-environ==0.9.0
django-filter==22.1
django-redis==5.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
environ==1.0
filelock==3.9.0
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
gcloud==0.18.3
geographiclib==2.0
geopy==2.3.0
googleapis-common-protos==1.58.0
greenlet==2.0.2
html5lib==1.1
httplib2==0.21.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jwcrypto==1.4.2
kombu==5.2.4
MarkupSafe==2.1.2
mixin==1.1
mixins==0.1.4
npm==0.1.1
oauth2client==4.1.3
optional-django==0.1.0
Pillow==9.4.0
platformdirs==3.0.0
prompt-toolkit==3.0.37
protobuf==4.22.0
psycopg2==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.17
pyfcm==1.5.4
PyJWT==2.6.0
pyparsing==3.0.9
Pyrebase4==4.6.0
python-dateutil==2.8.2
python-jwt==4.0.0
pytz==2022.7.1
razorpay==1.3.0
redis==4.5.1
requests==2.28.2
requests-toolbelt==0.10.1
rsa==4.9
six==1.16.0
soupsieve==2.4
SQLAlchemy==2.0.3
sqlparse==0.4.3
twilio==7.16.4
typing_extensions==4.4.0
tzdata==2022.7
urllib3==1.26.14
vine==5.0.0
virtualenv==20.19.0
wcwidth==0.2.6
webencodings==0.5.1
Werkzeug==2.2.2
wrapt==1.14.1
zope.interface==5.5.2